Screw -Wsign-conversion
Greatly simplified code as a result of not having to cast everyone and their
moms from unsigned to int and back, just to satisfy the compiler. Size is now
an int, as with all indices.

+++ src/sweeper.cpp
@@ -27,16 +27,16 @@ void Sweeper::tick(std::vector<sf::Event> &e, const sf::Time &t, sf::Vector2f m)
         m_elapsed += t.asSeconds();
 
     // process events
-    for (unsigned i = 0; i < e.size(); i++) {
+    for (int i = 0; i < (int)e.size(); i++) {
         if (e[i].type == sf::Event::MouseButtonPressed) {
             m_pressing = true;
             m_pressLoc = gAgent.getTile(e[i].mouseButton.x, e[i].mouseButton.y);
-            if (m_pressLoc.x == (unsigned)-1)
+            if (m_pressLoc.x == -1)
                 m_pressing = false;
         }
         else if (e[i].type == sf::Event::MouseButtonReleased) {
             if (m_pressing) {
-                sf::Vector2<unsigned> cur = gAgent.getTile(e[i].mouseButton.x, e[i].mouseButton.y);
+                sf::Vector2<int> cur = gAgent.getTile(e[i].mouseButton.x, e[i].mouseButton.y);
                 if (m_pressLoc.x == cur.x && m_pressLoc.y == cur.y)
                     m_status = mAgent.openTile(cur.y, cur.x);
                 m_pressing = false;
@@ -46,7 +46,7 @@ void Sweeper::tick(std::vector<sf::Event> &e, const sf::Time &t, sf::Vector2f m)
             m_pressing = false;
         }
     }
-    sf::Vector2<unsigned> mcur = gAgent.getTile(m.x, m.y);
+    sf::Vector2<int> mcur = gAgent.getTile(m.x, m.y);
     bool pressingInitLoc = m_pressing && m_pressLoc.x == mcur.x && m_pressLoc.y == mcur.y;
 
     gAgent.updateBoard(mAgent.getBoard(), m_status, pressingInitLoc, false, m_pressLoc);
