+++ src/sweepermechanics.cpp
@@ -15,21 +15,21 @@ void SweeperMechanics::reset()
     util::free2D(m_board);
     util::free2D(m_minefield);
     util::free2D(m_numfield);
-    m_size.x = (unsigned)config.getInt("board_width");
-    m_size.y = (unsigned)config.getInt("board_height");
+    m_size.x = config.getInt("board_width");
+    m_size.y = config.getInt("board_height");
     double mineProb = config.getInt("mine_perc") / 100.0;
     m_minefield = generateMinefield(mineProb, m_size.x, m_size.y);
     m_numfield = generateNumfield(m_minefield, m_size.x, m_size.y);
     m_board = util::alloc2Dint(m_size.x, m_size.y);
 }
 
-int SweeperMechanics::openTile(unsigned r, unsigned c)
+int SweeperMechanics::openTile(int r, int c)
 {
     if(m_board[r][c] == 0)
     {
         if(m_minefield[r][c])
         {
-            return 1; 
+            return 1;
         }
         else
         {
@@ -40,15 +40,15 @@ int SweeperMechanics::openTile(unsigned r, unsigned c)
     return 0; // game continue
 }
 
-int SweeperMechanics::triggerTile(unsigned r, unsigned c)
+int SweeperMechanics::triggerTile(int r, int c)
 {
     int flags = 0;
-    for(int i = (int)r - 1; i <= (int)r + 1; i++)
+    for(int i = r - 1; i <= r + 1; i++)
     {
-        for(int j = (int)c - 1; j <= (int)c + 1; j++)
+        for(int j = c - 1; j <= c + 1; j++)
         {
-            if(i >= 0 && i < (int)m_size.y
-              && j >= 0 && j < (int)m_size.x
+            if(i >= 0 && i < m_size.y
+              && j >= 0 && j < m_size.x
               && m_board[i][j] == 1)
             {
                 flags++;
@@ -58,32 +58,32 @@ int SweeperMechanics::triggerTile(unsigned r, unsigned c)
 
     if(flags == m_numfield[r][c])
     {
-        for(int i = (int)r - 1; i <= (int)r + 1; i++)
+        for(int i = r - 1; i <= r + 1; i++)
         {
-            for(int j = (int)c - 1; j <= (int)c + 1; j++)
+            for(int j = c - 1; j <= c + 1; j++)
             {
-                if(i >= 0 && i < (int)m_size.y
-                  && j >= 0 && j < (int)m_size.x
+                if(i >= 0 && i < m_size.y
+                  && j >= 0 && j < m_size.x
                   && m_board[i][j] == 0)
-                { 
-                    int res = openTile((unsigned)i, (unsigned)j);
+                {
+                    int res = openTile(i, j);
                     if(res == 1)
-                        return 1; 
+                        return 1;
                 }
             }
-        }               
+        }
     }
 
     return 0; // game continue
 }
 
-int SweeperMechanics::flagTile(unsigned r, unsigned c)
-{ 
+int SweeperMechanics::flagTile(int r, int c)
+{
     if(m_board[r][c] == 0)
         m_board[r][c] = 1;
     else if(m_board[r][c] == 1)
         m_board[r][c] = 0;
-    
+
     return 0;
 }
 
@@ -102,18 +102,18 @@ int** SweeperMechanics::getNumfield()
     return m_numfield;
 }
 
-sf::Vector2<unsigned> SweeperMechanics::getSize()
+sf::Vector2<int> SweeperMechanics::getSize()
 {
     return m_size;
 }
 
-bool** SweeperMechanics::generateMinefield(double p, unsigned w, unsigned h)
+bool** SweeperMechanics::generateMinefield(double p, int w, int h)
 {
     int pn = p * 100000;
     bool** minefield = util::alloc2Dbool(h, w);
-    for (unsigned i = 0; i < h; i++)
+    for (int i = 0; i < h; i++)
     {
-        for (unsigned j = 0; j < w; j++)
+        for (int j = 0; j < w; j++)
         {
             int r = rand() % 100000;
             minefield[i][j] = r < pn;
@@ -122,12 +122,12 @@ bool** SweeperMechanics::generateMinefield(double p, unsigned w, unsigned h)
     return minefield;
 }
 
-int** SweeperMechanics::generateNumfield(bool** minefield, unsigned w, unsigned h)
+int** SweeperMechanics::generateNumfield(bool** minefield, int w, int h)
 {
     int** numfield = util::alloc2Dint(h, w);
-    for (unsigned i = 0; i < h; i++)
+    for (int i = 0; i < h; i++)
     {
-        for (unsigned j = 0; j < w; j++)
+        for (int j = 0; j < w; j++)
         {
             if (i > 0)
             {
